Introducing the Scala library
The Scala library has just one dependency - the JDK. The implementation is a conversion from the Java one that we have in the typesafe-conductr repo.

A change has also been made to use the 307 status code instead of 308. We can go further though and handle some client side caching in the future if required (for the service locator in particular).