URI, URL and fallbacks
The JDK URI and URL types are now used instead of a String in order to reduce type conversions. In addition a fallback URI is now supplied to the service locator function in order to reduce the amount of boilerplate for a user of the API. A new ResourceOps files provides two convenience constructors for the JDK URI and URL types also - in order to keep the noise down.

Finally, the no-cache flavour of the locator lookup function was removed as the use of a cache is encouraged strongly.