Include releasing to maven into sbt release
So far the `sbt release` command was not closing and releasing the published artifacts on sonatype.

This PR is changing the `releaseProcess` task so that `sbt release` also releases the artifacts to sonatype.

The release process is now similar as for `constructr-zookeeper` in which this steps have been already battle tested: https://github.com/typesafehub/constructr-zookeeper/blob/master/project/Build.scala#L59-L73