Add sonatype release step
Adding `sonatypeReleaseAll` step after publishing the artifacts to sonatype. The `sonatypeProfileName` has been moved to `sonatype.sbt` so that sbt is picking up this value. The sonatype settings are not picked up in the `Build.scala`. If not specified, the sonatype profile name defaults to the organization name. This was the reason why the `sonatypeReleaseAll` command was not working prior to this PR.